+++ src/GJTransformControl.cpp
@@ -1,0 +1,17 @@
+#include "MoreObjectInfo.hpp"
+#include <Geode/binding/EditorUI.hpp>
+#include <Geode/modify/GJTransformControl.hpp>
+
+using namespace geode::prelude;
+using namespace std::string_view_literals;
+
+class $modify(MOITransformControl, GJTransformControl) {
+    static void onModify(ModifyBase<ModifyDerive<MOITransformControl, GJTransformControl>>& self) {
+        MoreObjectInfo::addHookListener(self, "GJTransformControl::ccTouchEnded", "dynamic-object-info"sv);
+    }
+
+    void ccTouchEnded(CCTouch* touch, CCEvent* event) {
+        GJTransformControl::ccTouchEnded(touch, event);
+        if (auto editorUI = EditorUI::get()) editorUI->updateObjectInfoLabel();
+    }
+};